Address Auguste review on dynamic metrics (#3174)
Inline callable-name checks, make _resolve_metric return None, simplify
docstrings/tests, restore setattr-based rmse monkeypatch, and move
dynamic metric tests to their own file.
