+++ examples/model_evaluation/plot_custom_metrics.py
@@ -75,9 +75,8 @@ report.metrics.get("specificity")
 
 # %%
 # Further, if the metric name is a valid identifier, the metric is also exposed as a
-# method — the same way built-in metrics such as ``accuracy`` or ``precision`` are:
+# method:
 report.metrics.specificity()
-report.metrics.accuracy()
 
 # %%
 # Passing extra keyword arguments
